_app/requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
brotli==1.0.9
django==4.2.3
djangorestframework==3.14.0
gunicorn==21.2.0
hiredis==2.2.3
packaging==23.1
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.7.1
whitenoise[brotli]==6.5.0
